Timers 0,1,2 should all work.
git-svn-id: https://fusion.wpi.edu/svn/rbe3001a15team2/trunk@30 015edbb1-928d-476b-afdc-d2170ecc5f68

+++ RBELib/include/RBELib/timer.h
@@ -46,6 +46,6 @@ void initTimer(int timer, int mode, unsigned int comp);
  *
  * @todo Create a function that will set a new compare value for the given timer.
  */
-void setCompValue(unsigned char timer, unsigned short int comp);
+void setCompValue(unsigned char timer, unsigned long int comp);
 
 #endif /* TIMER_H_ */
